pyproject.toml
[project]
name = "oparax-chirp"
version = "0.1.0"
description = "AI-powered social media automation for news reporters"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
